requirements.txt
openai
aiogram>=3.0.0
python-dotenv